pyproject.toml
[project]
name = "universaltoolkit"
description = "A comprehensive toolkit based on ComfyUI, providing image, mask, audio, and tools nodes, fully modular and v3 compatible."
version = "1.1.3"
license = {file = "LICENSE"}
dependencies = ["torch", "numpy", "Pillow", "opencv-python", "scipy", "tqdm"]
requires-python = ">=3.8"
classifiers = [
    "Operating System :: OS Independent",
    "Environment :: GPU :: NVIDIA CUDA"
]

[project.urls]
Repository = "https://github.com/whmc76/ComfyUI-UniversalToolkit"
Documentation = "https://github.com/whmc76/ComfyUI-UniversalToolkit/wiki"
"Bug Tracker" = "https://github.com/whmc76/ComfyUI-UniversalToolkit/issues"
#  Used by Comfy Registry https://registry.comfy.org

[tool.comfy]
PublisherId = "whmc76"
DisplayName = "ComfyUI-UniversalToolkit"
Icon = ""
Banner = ""
requires-comfyui = ">=1.0.0"
includes = []
